requirements.txt
git+https://github.com/facebookresearch/detectron2.git
gradio==3.16.2
albumentations==1.3.0
opencv-contrib-python
imageio==2.9.0
imageio-ffmpeg==0.4.2
pytorch-lightning==1.5.0
omegaconf==2.1.1
test-tube>=0.7.5
streamlit==1.12.1
einops==0.3.0
webdataset==0.2.5
kornia==0.6
open_clip_torch==2.0.2
invisible-watermark>=0.1.5
streamlit-drawable-canvas==0.8.0
torchmetrics==0.6.0
timm==0.6.12
addict==2.4.0
yapf==0.32.0
prettytable==3.6.0
safetensors==0.2.7
basicsr==1.4.2
numpy==1.23.1
fvcore
pycocotools
wandb
ipykernel
diffusers
pillow==9.0.1
accelerate==0.24.0
datasets
transformers==4.29.2
evaluate
huggingface_hub
wand
dask
opencv-python
wilds
scikit-image==0.20.0
git+https://github.com/openai/CLIP.git
xformers==0.0.22.post7
torch==2.1.0
torchvision==0.16.1
torchaudio==2.1.1